mcux: components: add specific files for connectivity support in Zephyr
IMU adapter and conn_fwloader need to be patched for Zephyr support.
Instead of patching them in mcux-sdk folder, it is preferred to
maintained them outside the sdk folder for the short term enablement.
It allows to keep the mcux-sdk folder as is.
On the long run, those patches should be backported to the official sdk
so they are available on future releases.

Those files are imported from feature/rw610_wifi_bringup branch based
on sha eb231289c10.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>