mcux: components: rpmsg: clear the rpmsg endpoint query response event
In the nxp_mfg_bridge sample, after ble initializes rpmsg, the rpmsg endpoint query response event is not cleared.
When 15.4 sends the rpmsg endpoint query, the event flag is not reset and does not wait on the k_event_wait(&rpmsgQFlagsRef) function.
And because the main task has a higher priority than the imu task, it cannot process the rpmsg endpoint query response,
so the enpoint query is sent all the time, causing the IMU fifo to be full.

Signed-off-by: Gang Li <gang.li_1@nxp.com>