mcux: hal_nxp: Include at least some parts of the cryptolib always
Those are required for startup. This essentially is the equivalent to

* 656d8543 (HEAD, tag: rw61x_ear_3.0) cmake:
Pull in els_pkc component [Declan Snyder] but with the option
to have the full cryptolib as a library (as is required for mbedtls
with acceleration and TF-M.
Updated cmake to include psa-crypto_driver lib include
paths and links the library always.

Signed-off-by: Stefan Krug <stefan.krug@nxp.com>
Signed-off-by: Waqar Ali Tahir <waqar.tahir@nxp.com>