mcux: components: psa_crypto_driver: Make driver buildable from TF-M build system.
If the psa_crypto_driver is used in a configuration that has TF-M, it is
built as part of TF-M S build. The TF-M S build is independent of Zephyr
therefore no Zephyr specific build flags can be used.
Passed on compiler options to both psa_crypto_driver and els_pkc cmake.

Signed-off-by: Stefan Krug <stefan.krug@nxp.com>
Signed-off-by: Waqar Tahir <waqar.tahir@nxp.com>