mcux: components: psa_crypto_driver: Fix key size calculation for opaque keys.
mcuxClPsaDriver_Oracle_GetKeyBufferSizeFromKeyData fills
key_buffer_size. It shall only be re-set to 0 in case the operation was
not sucessful.

We can not return the status of a mutex from this function, it is
expecting a length! There are no usages of ELS/PKC and no accesses to
global variables accessed in
mcuxClPsaDriver_Oracle_GetKeyBufferSizeFromKeyData, so there is no need
for mutexes in the first place.

Signed-off-by: Stefan Krug <stefan.krug@nxp.com>